local_store_memory_populator.py is a simple program that runs and periodically checks system memory and inserts into the local database.
app.py is a simple flask webserver. So far the only noun implemented is
memory_util.  I also implemented the "/?since=<seconds_since_epoch>" to
the right of memory_util.

psql_query.py (should be renamed local db query handler) is called when
a REST get is performed.  It accepts the "since" parameter and
converts the noun and parameter into a query. It returns the result.

json_producer.py takes a psql_query result and converts it into json
which app returns.

psql_insert (should be renamed to aggregator-fetcher or something)
calls the REST server (just localhost).  It gets the json result
parses it and inserts it into the aggregator db.

This is a rough implementation but works in the scenarios tested.
I'll test this in realtime streaming of mem utils tomorrow.