Here is how the database schemas will be generated automatically. The table manager reads the json file of event types and db_templates. The event types file is basically the file Chaos is creating for the data schema (not REST schema).
The keys of the event type dictionary each become a table.
Each event type in the event has a description
of the event, the units of type, the value column type for the schema.
Again this is like the data schema Chaos created.

It also provides which db_template to use.  So mem_used has a db
template of node, which has columns 'agg_id', 'node_id', 'ts'.
The 'v' columns can vary based on event type, hence why that column
type is in the event_types file.

All this is used to manage tables more easily and make it easier to
scale to many more tables.