Fixed an error with handling multiple ts filters.
For ts filters, the local store requires either
gt or gte and either lt or lte.

Small tweaks to the reference datastore related
to the lifetime of data in the local store.  The default
is 12 hours (in stats_populator.py).  For now the
lifetime of data in the local store could be a config
item.