(Forgot to add this in the last commit and failed at canceling the commit hence incomplete checkin comment)
I updated the validate_examples code based on our (Chaos and I)
better understanding of how UNIS and valedictory work.
Now we can make sure that extra keys in the json responses
from local datastores dont exist.  And the k-v pairs we receive
are of the proper type.

Added a node.schema file since we want to specify memory as
a key.

The schema files can now be pushed to the web.