I replaced the data_schema and event_schema files to be responses from the opsconfig local datastore.
For now the each program reads the example response from the
schema examples opsconfig directory.  When we standup an
opsconfig local datastore, the references will change and
point to that.

I've done all the unit-test and system tests I know to run.
The code passes all the tests.

I know I can refactor some of the tomorrow.  Most of the python
programs have a copy of a definition for loading the data_schema
and or the event_types.  This could be a new class in the common
directory.

Minor edits to the collector_operator.conf and alerting README