Added program to initialize a config local datastore.  This datastore publishes the list of monitoring events per object type.
I refactored the rest of the code to pull configuration from this
store.  If it cannot reach the store it reads it from the example
json response that would match the response from the config store.

I added a config item within the config directory for where the
config store is located.

The nuances with the way configuration is read in the table manager
caused me to edit so many files again.  I hope to be done editing
so many files per check in to reduce risk of error while we keep
improving our test infrastructure :)