Took a first cut at developing data plane pings values to the external check datastore.
The extckdatastore will run the coordinate_mesoscale_experiments.py
program to start pings from a VM in the mesoscale slice.

On the VM in the slice, there needs to be:
ip_addresses.conf
pinger.py
in the home directory
The pinger program is invoked by the
coordinate_mesoscale_experiments program.  It runs pings and outputs
raw sql statements into a file.
The coordinate program picks up this file, reads it in and puts it
into its database (which is the datastore for the external
check datastore)