Change the way the VMs are plumbed together to be simpler.
Before, some commands were run via puppet to attempt to find an IP
address on the host machine that the VMs would use to communicate.
Then, someone would have to manually copy the address to another file
to complete the configuration.  This was a) clumsy, and b) the
commands to find the IP address didn't work right.  (Somehow they
worked for Ryan, but I really don't see how.)

Now, we use puppet/virtualbox to configure host-only networking
on all of the VMs, and give them each a unique IP 192.168 address.
The collector VM can now contact the local datastore VM just fine
on the datastore's configured IP address, which should never change.

There is one bit of ugliness left because the datastore's IP address
appears in two places, but this is far better than before, so I'll
leave that for another day.