Remove references to starkville.bbn.com, Ryan's desktop machine.
starkville.bbn.com appeared in a URL to get opsconfig.json.  The
machine was powered off; when I powered it on, it had (not
surprisingly) an ancient version of opsconfig.json.  I replaced this
reference with a file URL to get opsconfig.json from the local
filesystem (Stephane's suggestion).  It's unclear if that's exactly
what we want in the long run, but it has to be better than starkville.