Check the value returned by handle_request before processing.
That value could be None if there was an issue during retrieval.
Also the "resources" and "slivers" array are optional in the aggregate info json response,
so test for their presence before processing them.

Finally modified the exception caught when issuing json.loads() and requests.get().
Catching only what we expect.

Addresses ticket:95