Better handling of various error path.
Looking at http error code (note: this was silently taken care of with
a if resp:, but in that case we don't know if we fail because we did not
get an response or we got an invalid response)

Also only catching known exceptions.

Took care of some warning issued by eclipse: unused import, unused variables, etc.

Addresses ticket:95