Use required field correctly.
In json schema draft-03, "required" is a boolean that is set
individually on each required field.  In draft-04, it is an array of
field names that are required.  Our schemas were using a mix of the
two styles, which is invalid in any case.  Change them all to the
draft-03 style (boolean).

With this change, our schemas pass validation according to
http://json-schema-validator.herokuapp.com/syntax.jsp .
(cherry picked from commit 3e2bd270acc2c319b35f21af21622b1ad7ba207a)