Remove incorrect notes about database schema changes.
I was under the mistaken impression that changes in the json schema
would automatically propagate to the database schema, but I guess
they have to be replicated in the db_schema entries in opsconfig.json.
As it turns out, the database schema was already using "interface"
everywhere, not "port," so this renaming will have no effect on the
tables.  Unless I'm wrong and still misunderstanding something.