Modified info_update() to not delete then insert data.
Now that we have foreign keys, data can not just be simply deleted if
some value from it is referenced in another table...
Using the new upsert() method of TableManager instead.

Added return value to info_update() and all methods using it,
and checking the calls.

exiting the script with a non-zero status if something went wrong during execution.

Addresses ticket:96.