Remove obsolete file.
Looking at the git history of this file and its peer gpo-ig.json,
it appears that Chaos and Ryan were working out how to represent the
top-level information from an aggregate.  This file was going to use
a UNIS schema, domain, directly, but that approach was abandonded n
favor of a custom schema for aggregates.