Set required_by_default to False for validictory.
In json schema-draft03, the "required" field states whether a property
is required to be present.  It is supposed to default to false, but
validictory overrides this and sets it to true, which means that all
fields are required by default.  This was causing validation of our
schemas to fail with validictory because of missing required fields.
Fortunately, validictory provides a way to override this behavior by
setting required_by_default=False in the validation call.  With this
change, all of our schemas pass validation using validictory.