Make some improvements suggested by Stéphane.
Now we just store the object id of the resource in the ops_sliver
table, and use that to look up the href/urn in either ops_node or
ops_link, instead of storing their href/urn redundantly.  Also, the
resource_type is deduced based on which table the id was found in
instead of needing an explicit field in the table.

Unfortunately, the nice idea of having node_id and link_id be
foreign keys gave me a lot of trouble, and I had to omit that.
The problem is that when you try to insert a NULL in a foreign
key field, mysql complains that it's invalid.  There is supposed
to be a way to have it accept NULLs in this case, but I tried
several options and none worked.

Addresses ticket:99