Use json.loads() instead of eval() to help parse the URL.
The part of the URL specifying what time-series data is being
requested really is a blob of JSON, so it's appropriate to use
json.loads here.  The JSON parser knows nothing about running Python
code, so this doesn't have the security issues that eval() does.

There was a second use of eval(), but the function it was in was dead
code, so I just removed the whole function.

Fixes ticket:126