Add software version to aggregate info page.
There are actually two versions.  monitoring_version is the version of
the GPO code and comes from a file named VERSION that make creates at
the top level of the repo using git.  populator_version is for the
rack aggregate teams to fill in a version for their code that
populates the database.  Its value comes from a new field in the
ops_aggregate table.

Addresses ticket:131