Update for new interface schema.
I had to fix the way info_update was handling insertion of NULL.  It
was inserting the string "NULL" instead of a bare SQL NULL.  The fact
that we were only ever inserting NULLs into string-typed columns (so
it never caused an error), combined with mysql printing a bare NULL
and the string "NULL" exactly the same in the table output, made this
quite a stealthy problem.

The string "NULL" is now special-cased to insert a bare NULL.  You'll
never be able to insert the actual string "NULL", but that's probably
a good thing.

I think this may have been causing problems with NULLs in foreign
key columns.  Up till now, we've never actually been inserting
a SQL NULL.

Addresses ticket:70
Addresses ticket:71