Eliminated redundant columns from ops_link_interfacevlan table.
urn and selfRef do not need to be in this table.

Also added comment about how to allow NULL values in the DB (which
up until now meant that the corresponding json fields were optional)
and yet report them. (use of a 4th boolean value in the DB schema)
In the end this functionality is not currently used, but leaving it
in there in case we need it again later.

Addresses ticket:84.