Corrected the handing of time series queries.
Also eliminated redundant methods: get_object_href_and_urn() and get_self_ref().
Using get_refs() instead.

Implemented method should_include_json_field(), instead of having the same test cut and pasted
all over the handle_XXX() methods.
Modified the logic of the test to force the reporting via json of a SQL NULL value. Until now and
for the most part, this meant that the json field is optional and therefore not going to be reported.
But a fourth (optional) field in the DB schema definition, can change that behavior to force the reporting of a NULL value.

This possibility ended up not being used for this ticket but leaving it in place in case we need it in the future.

Addresses ticket:84.