Corrected collector script to deal with DB constraints.
(at least for links and interface vlans)
Since we allow links to have a "remote" interfacevlan endpoint,
that endpoint ends up corresponding to an interface that is not related to a node,
hence any crawling of the nodes URLs will not get that interface.
The code now check that the interface correspondng to the interfacevlan exists
in the DB prior to inserting the interface vlans, and if not fetches it.

Added check_exists() method and moved some code to its own method:
refresh_interface_info(), so that it could be reused elsewhere.

Addresses ticket:84