Update for new release date.
The previous version of this script was somewhat fancier,
attempting to replace the date in the $schema lines of
the schemas with the correct date.  Now we just assume that
has been done, and copy the files to the web server as-is.

This task should probably be integrated into the makefile.