Modified DB schema for data tables related to external check stores.
Until now, there was an "aggregate_id" column that was added by default,
which was a foreign key in the ops_aggregate table (id column).
For the data tables whose records are coming from an external check store,
this was wrong, and is now an "externalcheck_id" column and a foreign key
to the ops_externalcheck (id column) table.

Addresses ticket:76.