Corrected the definition of the external check store VM.
Added the $datastore_ip_addr variable, and changed the apache installation
to be sslapache.

This VM is essentially the same as the local VM, and both the information from the
fake "local" datastore and the info from the external check store can be retrieved
from it. (the same is true for the local vagrant VM).
It's probably fine to remove that vagrant VM definition.

Addresses ticket:76.