Add --connection-close option.
This causes the HTTP request to have the "Connection: close" header so
that the server closes the connection immediately after sending the
request.  By default (without the option), we don't use this header.

This is more of a workaround than a fix, but it's useful to have for
now.

Addresses ticket:192