Adding the suspension of the VM after running the tests.
It seems that in the jenkins environment, jenkins kills whatever processes
it launched during the build, so the virtual box VM gets killed nastily.

Suspending it allows for the process to go away, yet takes a lot less time
to bring the VM up again on the next build...

Addresses ticket:31.