Corrected how a "stitcher" AM is registered.
an SCS AM section in opsconfig is going to contain possibly multiple amurl entries
(Any key starting with "amurl")
Also the am_nickname entry wil be used as the AM id in the DB.
Finally an am_status entry will tell if the SCS is production or development.

Addresses ticket:211.