Coded parallelization of running ping command at source sites.
Using a thread pool to execute all the commands related to running
pings at source sites.

For a given site, synching up the pinger script and IP configurations,
Executing the ping script, collecting the resulting file, inserting
the corresponding stats in the DB and deleting the file both locally
and remotely is executed by one of the threads of the thread pool.

So multiple soucre sites can be engaged simultaneously.

Addresses ticket:218.