Coded parallelization of pings.
Using thread pool to run the pings to the destinations.

For each destination, the pings "priming" the open flow rules,
the pings used to get the rtt value, and the parsing of the output
is executed by one of the thread pool.

Also made the ping count for "priming" the open flow rules, the ping count
for measuring the rtt and the thread pool size passed as arguments to the scripts.
(These are passed by the coordination script, which gets them from the extck configuration)

Addresses ticket:218.