Added new test for authority and externalcheck stores.
The object of the change was to test the new version field in the response
for authority info and externalcheck info. However these tests did not exist
until now.

Refactored the tests to have a base class with some common setup.
Added new TestExternalCheckStoreResponses and TestAuthorityStoreResponses test classes.
And of course, added the test for version info.

Addresses ticket:298.