Corrected SQL query for postgresql syntax.
in postgresql, since case is an issue, a case-sensitive column
name is quoted ("ColumnName") whereas in mysql it isn't (ColumnName).

However when prefixing the column name with the table name, the syntax is:
tablename."ColumnName" instead of "tablename.ColumnName"...

Also changed a bit the logic to avoid repeating the same queries about the same
metrics group.

NOTE: This would have been discovered earlier if the test fixture was testing both
flavors of DB.

Addresses ticket:309.